+++ models/villa.js
@@ -20,9 +20,11 @@ export default class VillaModel {
         }
     }
 
-    static async createVilla(body) {
-        const { name, description, address, capacity, price, images } = body
-        const validateResult = validateVilla(body)
+    static async createVilla(villaData) {
+        const { name, description, address, capacity, price, images } = villaData
+        const imagesJson = JSON.stringify(images || [])
+
+        const validateResult = validateVilla(villaData)
 
         if (!validateResult.success) {
             throw new Error(JSON.stringify(validateResult.error))
@@ -34,27 +36,20 @@ export default class VillaModel {
             }
 
             const safeDescription = description || null
-            const safeImages = images ? JSON.stringify(images) : null
 
             const [result] = await db.execute(
                 'INSERT INTO villas (name, description, address, capacity, price, images) VALUES (?, ?, ?, ?, ?, ?);',
-                [name, safeDescription, address, capacity, price, safeImages]
+                [name, safeDescription, address, capacity, price, imagesJson]
             )
-
-            if (result.affectedRows === 1) {
-                const [villas] = await db.execute('SELECT * FROM villas WHERE id = LAST_INSERT_ID();')
-                return villas[0]
-            } else {
-                throw new Error('Error creating villa')
-            }
+            return { id: result.insertId, ...villaData, images: images || [] }
         } catch (error) {
             console.log('Error creating villa:', error)
             throw error
         }
     }
 
-    static async updateVilla({ id, input }) {
-        const { name, description, address, capacity, price, images } = input
+    static async updateVilla({ id, villaData }) {
+        const { name, description, address, capacity, price, images } = villaData
         const updateFields = []
         const updatedValues = []
 
@@ -80,7 +75,7 @@ export default class VillaModel {
         }
         if (images !== undefined) {
             updateFields.push('images = ?')
-            updatedValues.push(JSON.stringify(images))
+            updatedValues.push(JSON.stringify(villaData.images))
         }
         if (updateFields.length === 0) {
             throw new Error('No valid fields to update')
@@ -91,7 +86,11 @@ export default class VillaModel {
             const [result] = await db.execute(`UPDATE villas SET ${updateFields.join(', ')} WHERE id = ?;`, updatedValues)
 
             if (result.affectedRows > 0) {
-                return { message: 'Villa updated successfully' }
+                const [updatedVilla] = await db.execute('SELECT * FROM villas WHERE id = ?;', [id])
+                if (updatedVilla[0].images !== null && updatedVilla[0].images !== undefined) {
+                    updatedVilla[0].images = updatedVilla[0].images
+                }
+                return updatedVilla[0]
             } else {
                 return { message: 'Villa not found' }
             }
@@ -103,6 +102,11 @@ export default class VillaModel {
 
     static async deleteVilla(id) {
         try {
+            const [car] = await db.execute('SELECT images FROM villas WHERE id = ?;', [id])
+            if (car.length === 0) {
+                return { success: false, message: 'Villa not found' }
+            }
+
             const [result] = await db.execute('DELETE FROM villas WHERE id = ?;', [id])
 
             if (result.affectedRows > 0) {
